:recycle: Centralize types in index.js
