+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "@frenchpastries/millefeuille",
   "version": "0.1.6",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "index.js",
+  "types": "index.d.ts",
   "repository": "https://github.com/FrenchPastries/millefeuille",
   "author": "Guillaume Hivert <guillaume.hivert@outlook.com>",
   "license": "MIT",
